packages/python-backend/requirements.txt
# Revelio Python Backend Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn>=0.27.0

# AI/ML
google-genai>=0.3.0
fal-client>=0.4.0
tavily-python>=0.3.0

# Image Processing
Pillow>=10.0.0

# Data/Utils
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
feedparser>=6.0.0

# Search (fallback)
duckduckgo-search>=4.0.0

# File uploads
python-multipart>=0.0.6
